requirements.txt
google-cloud-texttospeech
google-generativeai